fix comment bug improve close modal

+++ app/controllers/api/hashtags_controller.rb
@@ -8,18 +8,17 @@ class Api::HashtagsController < ApplicationController
   end
 
   def create
-    # debugger
     post_id = hashtags_params[:post_id]
     @relationships = {post_id: post_id, hashtag_id_array: []}
-    hashtags_params[:hashtags_array].each do |pair|
+    hashtags_params[:hashtags_array].uniq.each do |hashtag|
       # hashtags_params[:hashtags_array] looks like { 0 => [#newyork, 2], 1 => [#summer, 9] }
-      if existing_hashtag = Hashtag.find_by(hashtag: pair[1][0])
+      if existing_hashtag = Hashtag.find_by(hashtag: hashtag )
         existing_hashtag.update(count: existing_hashtag.count += 1)
         @relationships[:hashtag_id_array].push(existing_hashtag.id)
       else
-        hashtag = Hashtag.new(hashtag: pair[1][0], count: 1)
-        if hashtag.save
-          @relationships[:hashtag_id_array].push(hashtag.id)
+        new_hashtag = Hashtag.new(hashtag: hashtag, count: 1)
+        if new_hashtag.save
+          @relationships[:hashtag_id_array].push(new_hashtag.id)
         end
       end
     end
@@ -36,10 +35,9 @@ class Api::HashtagsController < ApplicationController
   def destroy_many_hashtags
     delete_many_params.each do |hashtag_id|
       hashtag = Hashtag.find(hashtag_id)
-      if hashtag.count == 1
-        # debugger
+      if hashtag && hashtag.count == 1
         hashtag.destroy
-      else
+      elsif hashtag
         new_count = hashtag.count - 1
         hashtag.update(count: new_count)
       end
@@ -61,7 +59,9 @@ class Api::HashtagsController < ApplicationController
   def hashtags_params
     # hashtags_params = params.require(:hashtags).permit(:post_id).permit(:hashtags_array)
     hashtags_params = params.require(:hashtags).permit(:post_id)
-    hashtags_params[:hashtags_array] = params[:hashtags][:hashtags_array]
+    hashtags_params[:hashtags_array] = params[:hashtags][:hashtags_array].map do |idx, hashtag|
+      hashtag[0].downcase
+    end
     hashtags_params
   end
 end
